Move memberships test in keeper package
