+++ x/memberships/genesis_test.go
@@ -3,7 +3,11 @@ package memberships_test
 import (
 	"testing"
 
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	"github.com/commercionetwork/commercionetwork/x/memberships"
+	"github.com/commercionetwork/commercionetwork/x/memberships/keeper"
+	"github.com/commercionetwork/commercionetwork/x/memberships/types"
 
 	"github.com/stretchr/testify/require"
 )
@@ -14,19 +18,18 @@ func TestDefaultGenesisState(t *testing.T) {
 }
 
 func TestInitGenesis(t *testing.T) {
-	/*defGen := memberships.DefaultGenesisState("uccc")
-	ctx, _, _, k := SetupTestInput()
-	require.Equal(t, memberships.GenesisState{LiquidityPoolAmount: sdk.Coins(nil), Invites: types.Invites{}, TrustedServiceProviders: ctypes.Addresses{}, Memberships: types.Memberships{}}, defGen)
-	memberships.InitGenesis(ctx, k, k.SupplyKeeper, defGen)*/
-	/*export := memberships.ExportGenesis(ctx, k)
-	require.Equal(t, commerciomint.GenesisState{Positions: []types.Position{}, LiquidityPoolAmount: sdk.Coins(nil), CreditsDenom: "test", CollateralRate: sdk.NewDec(2)}, export)
-	*/
+	defGen := memberships.DefaultGenesisState("uccc")
+	ctx, _, _, k := keeper.SetupTestInput()
+	require.Equal(t, memberships.GenesisState{LiquidityPoolAmount: sdk.Coins(nil), Invites: nil, TrustedServiceProviders: nil, Memberships: nil}, defGen)
+	memberships.InitGenesis(ctx, k, k.SupplyKeeper, defGen)
+	export := memberships.ExportGenesis(ctx, k)
+	require.Equal(t, memberships.GenesisState{LiquidityPoolAmount: sdk.Coins(nil), Invites: types.Invites{}, TrustedServiceProviders: nil, Memberships: types.Memberships{}}, export)
 }
 
-func TestExportGenesis(t *testing.T) {
+/*func TestExportGenesis(t *testing.T) {
 
 }
 
 func TestValidateGenesis(t *testing.T) {
 
-}
+}*/
