+++ x/memberships/keeper/keeper_invites_test.go
@@ -1,4 +1,4 @@
-package keeper_test
+package keeper
 
 import (
 	"fmt"
