+++ x/memberships/keeper/keeper_rewards_test.go
@@ -1,4 +1,4 @@
-package keeper_test
+package keeper
 
 import (
 	"testing"
